+++ CHANGELOG.md
@@ -11,11 +11,17 @@
     - Write POST request data properly to file when ran with `-od`.
     - Fixed a bug by using header canonicaliztion related to HTTP headers being case insensitive.
 <<<<<<< HEAD
+<<<<<<< HEAD
     - Properly handle relative redirect urls with `-recursion`
     - Calculate req/sec correctly for when using recursion
     - When `-request` is used, allow the user to override URL using `-u`
 =======
 >>>>>>> formatting, canonical customer headers, docs updated
+=======
+    - Properly handle relative redirect urls with `-recursion`
+    - Calculate req/sec correctly for when using recursion
+    - When `-request` is used, allow the user to override URL using `-u`
+>>>>>>> moved logic back and less resolve conflicts
 
 - v1.0.1
   - Changed